LPS-151267 search-experiences-service:separate upgrade processes
https://issues.liferay.com/browse/LPS-151267
