+++ modules/dxp/apps/search-experiences/search-experiences-service/bnd.bnd
@@ -2,6 +2,6 @@ Bundle-Name: Liferay Search Experiences Service
 Bundle-SymbolicName: com.liferay.search.experiences.service
 Bundle-Version: 3.0.1
 Liferay-Enterprise-App: dxp.only=true;product.id=22b7e30f-34d4-4a63-9696-56987ad66e4e
-Liferay-Require-SchemaVersion: 1.1.0
+Liferay-Require-SchemaVersion: 1.2.0
 Liferay-Service: true
 -dsannotations-options: inherit
